requirements.txt
# =========================
# CORE
# =========================
numpy==1.26.4
pandas==2.2.2

# =========================
# MACHINE LEARNING
# =========================
scikit-learn==1.5.0
scikit-image==0.22.0
mahotas==1.4.13

# =========================
# COMPUTER VISION
# =========================
opencv-python==4.9.0.80
Pillow==12.2.0

# =========================
# VISUALIZAÇÃO
# =========================
plotly==5.22.0
streamlit==1.37.0

# =========================
# BANCO DE DADOS
# =========================
SQLAlchemy==2.0.30

# =========================
# EXPORTAÇÃO
# =========================
openpyxl==3.1.2
reportlab==4.2.0

# =========================
# UTIL
# =========================
joblib==1.4.2